Read from stdout instead of stderr
If stdout is connected to a terminal but stderr isn't, it will seem like
lots is ignoring input, because it will still enter its "interactive"
mode. However, if stdout is not connected to a terminal, it will enter
its "cat-like" or "non-interactive" mode, not requiring input anyway.

By using stdout for both input and output, interactiveness and input
availability will be consistent.